socketdir: stop proxying when backend (half-)closes
This is necessary to avoid leaving connections open indefinitely
if the backend closes and the client never sends more data.

The downside is that if the backend only half-closes we will stop proxying
immediately instead of continuing to proxy data from the client.  Ideally,
we would only stop once we knew for sure the backend connection had
been broken.  On Linux, this can be detected by enabling TCP keepalives
and polling the socket with events=0 until revents contains POLLHUP.
Unfortunately, Go's standard library doesn't expose POLLHUP events.
Until it does, we will not be able to support proxying to backends which
use half-closed connections, but this seems pretty unusual.