cert: allow PEM header to specify private key usage (experimental)
You can add "Usage: sign" or "Usage: decrypt" to the header of a
private key's PEM block to limit the key to just that usage.  This
is useful for restricting an RSA key to signing only, effectively
disabling RSA key exchange.

This is experimental and may be changed or removed.